backend/requirements.txt
Django==4.1.7
django-filter==23.1
djoser==2.2.0
Pillow==9.5.0
python-dotenv==1.0.0
flake8==6.0.0
pdfkit==1.0.0